use inline style
